Refactor TextInput
